IJPS-36 Chnge to folder struct,add role to cookie
