+++ api/auth/index.ts
@@ -5,6 +5,7 @@ import { getConfigValue, showFeature } from '../configuration';
 import {
   COOKIES_TOKEN,
   COOKIES_USER_ID,
+  COOKIES_USER_ROLE,
   FEATURE_SECURE_COOKIE_ENABLED,
   IDAM_SECRET,
   LOGIN_ROLE_MATCHER,
@@ -26,10 +27,11 @@ export const successCallback = (req: EnhancedRequest, res: Response, next: NextF
   const {accessToken} = user.tokenset;
   const cookieToken = getConfigValue(COOKIES_TOKEN);
   const cookieUserId = getConfigValue(COOKIES_USER_ID);
+  const cookieUserRole = getConfigValue(COOKIES_USER_ROLE);
 
   res.cookie(cookieUserId, userinfo.uid);
   res.cookie(cookieToken, accessToken);
-  res.cookie('__role__', userinfo.roles);
+  res.cookie(cookieUserRole, userinfo.roles);
   if (!req.isRefresh) {
     return res.redirect('/');
   }
