+++ config/default.json
@@ -2,7 +2,8 @@
     "environment": "production",
     "cookies": {
         "token": "__auth__",
-        "userId": "__userid__"
+        "userId": "__userid__",
+        "userRole": "__userrole__"
     },
     "now": false,
     "services": {
