+++ src/app/_guards/auth/auth.guard.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { SessionStorageService } from '../session-storage/session-storage.service';
-import { WindowLocationService } from '../window-location/window-location.service';
-import { AuthService } from './auth.service';
+import { SessionStorageService } from '../../_services/session-storage/session-storage.service'
+import { WindowLocationService } from '../../_services/window-location/window-location.service';
+import { AuthService } from '../../_services/auth/auth.service'
 
 
 @Injectable({
