--- src/app/_services/venue.service.ts
@@ -1,26 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { VenueModel } from '../_models/venue.model';
-import { Observable } from 'rxjs';
-import { CookieService } from 'ngx-cookie-service';
-@Injectable({
-  providedIn: 'root'
-})
-export class VenueService {
-
-  constructor(
-    private readonly http: HttpClient,
-    private readonly cookies: CookieService
-  ) { }
-
-  public getAllVenues(searchTerm: string): Observable<VenueModel[]> {
-
-    // TODO: move these to service file
-    const S2SToken: string  = 'M2GT4N2JJA6YEUA5';
-    const authToken: string = this.cookies.get('__auth__')
-    
-    return this.http.post<VenueModel[]>('/test', {S2S: S2SToken, AUTH: authToken, SEARCHSTRING: searchTerm});
-
-  }
-
-}
