+++ src/app/_services/window-location/window-location.service.ts
