+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ApplicationRoutingComponent } from './components/routing/application-routing.component';
-import { AuthGuard } from './services/auth/auth.guard';
+import { AuthGuard } from './_guards/auth/auth.guard';
 
 const routes: Routes = [
   {
