+++ src/app/sitting-records/sitting-records-routing.module.ts
@@ -4,8 +4,8 @@ import { RouterModule, Routes } from '@angular/router';
 import { SittingRecordsComponent } from './sitting-records.component';
 import { ManageSittingRecordsComponent } from './manage-sitting-records/manage-sitting-records.component';
 import { ViewSittingRecordsComponent } from './view-sitting-records/view-sitting-records.component';
-import { SittingRecordsGuard } from '../_guards/sitting-records.guard';
-import { AuthGuard } from '../services/auth/auth.guard';
+import { SittingRecordsGuard } from '../_guards/sitting-records/sitting-records.guard';
+import { AuthGuard } from '../_guards/auth/auth.guard';
 
 const routes: Routes = [
     {
